backend/requirements.txt
# Backend Requirements - Hot and Cold Game API

# FastAPI framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic[email]>=2.0.0

# Machine Learning / NLP
gensim>=4.3.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Database
sqlalchemy>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# File handling
python-multipart>=0.0.6
pillow>=10.0.0  # Image processing for avatar resize

# Utilities
requests>=2.31.0  # Per test script
